SchoolJavaEveris-Stock_Produ

+++ escuela-productoMS/src/main/java/com/example/demo/dto/CantidadDTO.java
@@ -12,6 +12,6 @@ import lombok.Setter;
 public class CantidadDTO {
 
 	@JsonProperty(value = "cantidad")
-	private Long cantidadStock;
+	private int cantidadStock;
 
 }
