+++ escuela-productoMS/src/main/java/com/example/demo/dto/ProductoDTO.java
@@ -26,6 +26,7 @@ public class ProductoDTO {
 	private ImagenProducto imagenProducto;
 	
 	//CANTIDAD
-	private Long cantidadStock;
+	private int cantidadStock;
+
 
 }
