+++ escuela-stockMS/src/main/java/com/example/demo/entity/Stock.java
@@ -28,6 +28,6 @@ public class Stock {
 	private Long idTienda;
 	
 	@Column
-	private int cantidad;
+	private Long cantidad;
 	
 }
